update workspace functions
